formatter: remove broken vale configuration
Removes the broken and uncommented value configuration for vale.

While vale can be a valuable tool, we need a working configuration if we
want to use it. Keeping a broken configuration on main is not a
sustainable model.

We can fix it and reintroduce it cleanly at a later stage.