fix(installer): Fix language check case
When attempting to follow the docs on the installer, this error comes
up:
```
$ clan flash --flake git+https://git.clan.lol/clan/clan-core \
             --ssh-pubkey ~/.ssh/id_ed25519.pub \
             --keymap us \
             --language en_US.utf-8 \
             --dry-run flash-installer

ERROR: Language 'en_US.utf-8' is not a valid language. Run 'clan flash --list-languages' to see a list of possible languages.
```

I only decided to clean up the docs, but I'm also okay with just
downcasing the locales in `list_possible_languages()`.